chore(debloat): execute E1-E5 dispositions — delete findings/compute/hook-event-writer, re-point E2/E3 reservations (#1713)
- E1 review/findings.ts: DELETED with findings.test.ts. The review.finding /
  review.escalated schema registrations and their projection consumers stay
  (schema is contract, not emitter); both events are now emitter-less.
- E2 benchmarks/baselines-schema.ts: KEPT dormant; RESERVED re-pointed from
  #1713 to the benchmark-validation epic #1677 (expiry 2026-10-31 unchanged).
  Not moved into the benchmark-harness allowlist class — that class excludes
  *-schema.ts contract surfaces by design.
- E3 adapters/remote-mcp.ts: KEPT reserved; RESERVED header now cites #1081
  (DR-6 remote-MCP design) only, interim #1713 tracking tail removed (expiry
  2026-10-31 unchanged).
- E4 runbooks/compute.ts: DELETED; its two test callers in runbooks/drift.test.ts
  (compute-vs-declared autoEmits cross-checks) removed with it. The
  registry-side RunbookDrift_AutoEmitsMatchEventEmissionRegistry check remains.
- E5 event-store/hook-event-writer.ts: DELETED with hook-event-writer.test.ts.
  storage/sidecar-merger.test.ts now writes raw sidecar JSONL via a local
  helper (same pattern sidecar-scheduler.test.ts already used), preserving
  the merger's wire-format coverage.

Appended a §6 escalation-outcomes note to the 005 disposition baseline spec.
Gates verified locally: check-module-intent exit 0, knip-diff exit 0
(55 findings, all allowlisted), server typecheck clean, scoped suites green.
