--- servers/exarchos-mcp/src/review/findings.ts
@@ -1,67 +1,0 @@
-// ─── Review Finding & Escalation Event Emission ─────────────────────────────
-//
-// Utility functions for emitting review.finding and review.escalated events.
-// These can be called when review processing (e.g., CodeRabbit comment parsing
-// or review triage) identifies actionable findings or escalation conditions.
-// ────────────────────────────────────────────────────────────────────────────
-
-// RESERVED(issue: #1713, owner: exarchos, expires: 2026-10-31) — cascade orphan; emits registered review.finding/review.escalated, named as the utility to wire by 2 design docs; delete-or-wire tracked in #1713 (DR-7 module-intent gate)
-
-import type { EventStore } from '../event-store/store.js';
-import type { ReviewFinding, ReviewEscalated } from '../event-store/schemas.js';
-
-/**
- * Emit a `review.finding` event for each actionable finding.
- *
- * Fire-and-forget: individual append failures are silently swallowed
- * so callers are never blocked by event emission errors.
- */
-export async function emitReviewFindings(
-  findings: ReviewFinding[],
-  streamId: string,
-  eventStore: EventStore,
-): Promise<void> {
-  for (const finding of findings) {
-    try {
-      await eventStore.append(streamId, {
-        type: 'review.finding',
-        data: {
-          pr: finding.pr,
-          source: finding.source,
-          severity: finding.severity,
-          filePath: finding.filePath,
-          lineRange: finding.lineRange,
-          message: finding.message,
-          rule: finding.rule,
-        },
-      });
-    } catch {
-      // Intentionally swallowed — event emission is fire-and-forget
-    }
-  }
-}
-
-/**
- * Emit a `review.escalated` event when a PR is escalated to human review.
- *
- * Fire-and-forget: append failures are silently swallowed.
- */
-export async function emitReviewEscalated(
-  escalation: ReviewEscalated,
-  streamId: string,
-  eventStore: EventStore,
-): Promise<void> {
-  try {
-    await eventStore.append(streamId, {
-      type: 'review.escalated',
-      data: {
-        pr: escalation.pr,
-        reason: escalation.reason,
-        originalScore: escalation.originalScore,
-        triggeringFinding: escalation.triggeringFinding,
-      },
-    });
-  } catch {
-    // Intentionally swallowed — event emission is fire-and-forget
-  }
-}
